+++ src/helpers/BaseHelper/form.helper.ts
@@ -126,6 +126,7 @@ export class FormHelper extends BaseHelper {
                         name,
 
                         option: value ? String(value) : '',
+                        exact: true,
                     });
 
                     break;
